Add /public/states/info API call
- Break out the public API into its own file
- Add additional columns to State model to match public API
