+++ app/api/data.py
@@ -34,18 +34,6 @@ def get_batches():
 
 
 ##############################################################################################
-######################################   States      #########################################
-##############################################################################################
-
-@api.route('/states', methods=['GET'])
-def get_states():
-    current_app.logger.info('Retrieving all states')
-    states = State.query.all()
-    return jsonify({
-        'states': [state.to_dict() for state in states]
-    })
-
-##############################################################################################
 ######################################   Core data      ######################################
 ##############################################################################################
